StorePodcastScreen : new page - Use PodcastArg as navigation argument - Show first / more button / last episode (Castro inspiration) - Remove PagerMediator - Add related podcasts * Update database SQL and related repositories

+++ androidApp/src/main/java/com/caldeirasoft/outcast/ui/util/DurationFormatter.kt
@@ -1,10 +1,11 @@
 package com.caldeirasoft.outcast.ui.util
 
+import java.lang.StringBuilder
 import kotlin.time.DurationUnit
 import kotlin.time.toDuration
 
 object DurationFormatter {
-    fun Long.formatDuration(): String {
+    fun Int.formatDuration(): String {
         val duration = this.toDuration(DurationUnit.SECONDS)
         return duration.toComponents { hours, minutes, seconds, nanoseconds ->
             with(StringBuilder()) {
